requirements.txt
deta
streamlit_extras
python-docx
datetime
pytz
streamlit_js_eval